requirements.txt
django
youtube-search-python
googlesearch-python
nltk
spacy